configure: add configure script to handle fmemopen
Previously two commits added a fallback fmemopen for OS X/BSD that
didn't support fmemopen, then fixed some of the ifdef logic around it.
Unforunately the logic was backwards, as libc implementations do not
define _POSIX_C_SOURCE in order to advertise features.  Instead the
user defines _POSIX_C_SOURCE to control which definitions are
available.

At first glance _POSIX_VERSION could be tested to see if fmemopen
is available, but this only solves half the problem.  The fallback
implementation uses then non-standard BSD specific funopen.  This means
that even if _POSIX_VERSION is tested, there needs to be a way to
test for the existence of funopen.

Solve the problem with a short configure script that checks for
fmemopen and funopen.  This will work independent of which libc and
version are used.

Fixes: 45c7237 (Fixup for recent firmware inclusion changes (#34))
Fixes: 17f6b01 (Fix cross-platform building)